Adding option for initial highlight. Viewer can be initialized with a highlight active. Also consolidated highlight code into render functions.
