+++ src/RenderFunctions.js
@@ -156,5 +156,52 @@ function getTranslate(transform)
       // As per definition values e and f are the ones for the translation.
       return [matrix.e, matrix.f];
   }
+  function setHighlights(selectedAlleles,svgTarget){
+    let viewer_height= svgTarget.node().getBBox().height;
+    let highlights=svgTarget.selectAll(".variant-deletion,.variant-SNV,.variant-insertion,.variant-delins")
+      .filter(function(d){
+        let returnVal=false;
+        //TODO This needs to be standardized.  We sometimes get these returned in a comma sperated list
+        //and sometimes in an array.
+        if(d.alleles){
+          let ids=d.alleles[0].replace(/"|\[|\]| /g, "").split(',');
+          ids.forEach((val) => {
+            if (selectedAlleles.includes(val)){
+              returnVal=true;
+            }
+          })
+          d.alleles.forEach((val) => {
+            if (selectedAlleles.includes(val)){
+              returnVal=true;
+            }
+          })
+        }
+        return returnVal;
+      })
+      .datum(function(d){
+        d.selected="true";
+        return d;
+      })
+      .style("stroke" , "black")
 
-export { countIsoforms, findRange, checkSpace, doResize, calculateNewTrackPosition, getTranslate  }
+      highlights.each(function(){
+        let x_val=d3.select(this).attr('x');
+        let width_val=d3.select(this).attr('width');
+        if(width_val == null){
+          width_val = 3;
+          x_val = x_val-(width_val/2);
+        }
+        svgTarget.select(".deletions.track")
+        .append('rect')
+        .attr("class","highlight")
+        .attr("x", x_val)
+        .attr("width", width_val)
+        .attr("height", viewer_height)
+        .attr("fill", 'yellow')
+        .attr("opacity", 0.8)
+        .lower();
+      })
+  }
+
+
+export { countIsoforms, findRange, checkSpace, doResize, calculateNewTrackPosition, getTranslate, setHighlights  }
