+++ src/demo/index.js
@@ -28,9 +28,9 @@ function covidExamples(){
 }
 
 function currentExamples(){
-  createExample("2R:11524006..11528210", "fly", "viewerActnFly", TRACK_TYPE.ISOFORM_AND_VARIANT, false,[],[]);
+  createExample("2R:11524006..11528210", "fly", "viewerActnFly", TRACK_TYPE.ISOFORM_AND_VARIANT, false,[],[],['FB:FBal0299736','FB:FBal0003773']);
   //createExample("8:57320983..57324517", "mouse", "viewerHand2Mouse", TRACK_TYPE.ISOFORM_AND_VARIANT, false,[],['ENSMUST00000185635']);
-  //createExample("9:42732992..42873700", "zebrafish", "viewerHighlightExample", TRACK_TYPE.ISOFORM_AND_VARIANT, false);
+  createExample("9:42732992..42873700", "zebrafish", "viewerHighlightExample", TRACK_TYPE.ISOFORM_AND_VARIANT, false);
   //createCoVExample("NC_045512.2:17894..28259", "SARS-CoV-2", "covidExample1", TRACK_TYPE.ISOFORM, false);
   //createHTPExample("X:2023822..2042311", "fly", "viewerActnHTPFly", TRACK_TYPE.ISOFORM, false,[],'Actn','X:2037135');
 }
@@ -99,7 +99,7 @@ function isoformExamples() {
     createIsoformExample("25:15029041..15049781", "zebrafish", "zebrafishExampleIsoformOnly", TRACK_TYPE.ISOFORM, true);
 }
 
-function createExample(range, genome, divId, type, showLabel, variantFilter,isoformFilter) {
+function createExample(range, genome, divId, type, showLabel, variantFilter,isoformFilter,initialHighlight) {
     const chromosome = range.split(":")[0];
     const [start, end] = range.split(":")[1].split("..");
     const ratio = 0.01;
@@ -112,6 +112,7 @@ function createExample(range, genome, divId, type, showLabel, variantFilter,isof
       "transcriptTypes": getTranscriptTypes(),
       "isoformFilter": isoformFilter || [],
       "variantFilter": variantFilter || [],
+      "initialHighlight" : initialHighlight || [],
         "binRatio": ratio,
         "tracks": [
             {
